[#105] #REVIEW 'assemblyName: DotNet.Testcontainers.Tests; refactor: DatabaseContainerTest' {Move database tests into separate classes.}
