+++ src/DotNet.Testcontainers.Tests/Unit/Linux/Database/MySqlTestcontainerTest.cs
@@ -1,0 +1,44 @@
+namespace DotNet.Testcontainers.Tests.Unit.Linux.Database
+{
+  using System.Threading.Tasks;
+  using DotNet.Testcontainers.Core.Builder;
+  using DotNet.Testcontainers.Core.Containers.Database;
+  using DotNet.Testcontainers.Core.Models.Database;
+  using MySql.Data.MySqlClient;
+  using Xunit;
+
+  public class MySqlTestcontainerTest
+  {
+    [Fact]
+    public async Task ConnectionEstablished()
+    {
+      // Given
+      var testcontainersBuilder = new TestcontainersBuilder<MySqlTestcontainer>()
+        .WithDatabase(new MySqlTestcontainerConfiguration
+        {
+          Database = "db",
+          Username = "mysql",
+          Password = "mysql",
+        });
+
+      // When
+      // Then
+      using (var testcontainer = testcontainersBuilder.Build())
+      {
+        await testcontainer.StartAsync();
+
+        using (var connection = new MySqlConnection(testcontainer.ConnectionString))
+        {
+          connection.Open();
+
+          using (var cmd = new MySqlCommand())
+          {
+            cmd.Connection = connection;
+            cmd.CommandText = "SELECT 1";
+            cmd.ExecuteReader();
+          }
+        }
+      }
+    }
+  }
+}
