+++ src/DotNet.Testcontainers.Tests/Unit/Linux/Database/PostgreSqlTestcontainerTest.cs
@@ -1,0 +1,44 @@
+namespace DotNet.Testcontainers.Tests.Unit.Linux.Database
+{
+  using System.Threading.Tasks;
+  using DotNet.Testcontainers.Core.Builder;
+  using DotNet.Testcontainers.Core.Containers.Database;
+  using DotNet.Testcontainers.Core.Models.Database;
+  using Npgsql;
+  using Xunit;
+
+  public class PostgreSqlTestcontainerTest
+  {
+    [Fact]
+    public async Task ConnectionEstablished()
+    {
+      // Given
+      var testcontainersBuilder = new TestcontainersBuilder<PostgreSqlTestcontainer>()
+        .WithDatabase(new PostgreSqlTestcontainerConfiguration
+        {
+          Database = "db",
+          Username = "postgres",
+          Password = "postgres",
+        });
+
+      // When
+      // Then
+      using (var testcontainer = testcontainersBuilder.Build())
+      {
+        await testcontainer.StartAsync();
+
+        using (var connection = new NpgsqlConnection(testcontainer.ConnectionString))
+        {
+          connection.Open();
+
+          using (var cmd = new NpgsqlCommand())
+          {
+            cmd.Connection = connection;
+            cmd.CommandText = "SELECT 1";
+            cmd.ExecuteReader();
+          }
+        }
+      }
+    }
+  }
+}
