+++ tools/packages.config
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <packages>
-    <package id="Cake" version="0.32.1" />
+  <package id="Cake" version="0.33.0" />
 </packages>
